Refactor schemaGenerator folder structure
