+++ src/generate/SchemaError.ts
@@ -1,0 +1,11 @@
+// @schemaDefinition: A GraphQL type schema in shorthand
+// @resolvers: Definitions for resolvers to be merged with schema
+export default class SchemaError extends Error {
+  public message: string;
+
+  constructor(message: string) {
+    super(message);
+    this.message = message;
+    Error.captureStackTrace(this, this.constructor);
+  }
+}
